+++ tests/test_download.py
@@ -226,3 +226,83 @@ def test_download_invalid_id(api):
     with pytest.raises(SentinelAPIError) as excinfo:
         api.download(uuid)
     assert "Invalid key" in excinfo.value.msg
+
+
+@pytest.mark.vcr
+@pytest.mark.scihub
+def test_download_quicklook(api, tmpdir, quicklook_products):
+    uuid = quicklook_products[0]["id"]
+    filename = quicklook_products[0]["title"]
+    expected_path = tmpdir.join(filename + ".jpeg")
+    # import pdb; pdb.set_trace()
+    # Download normally
+    quicklook_info = api.download_quicklook(uuid, str(tmpdir))
+    assert expected_path.samefile(quicklook_info["path"])
+    assert quicklook_info["title"] == filename
+    assert quicklook_info["quicklook_size"] == expected_path.size()
+    assert quicklook_info["downloaded_bytes"] == expected_path.size()
+
+    modification_time = expected_path.mtime()
+    expected_quicklook_info = quicklook_info
+
+    # File exists, expect nothing to happen
+    quicklook_info = api.download_quicklook(uuid, str(tmpdir))
+    assert expected_path.mtime() == modification_time
+    expected_quicklook_info["downloaded_bytes"] = 0
+    assert quicklook_info["quicklook_size"] == expected_path.size()
+    assert quicklook_info == expected_quicklook_info
+
+    tmpdir.remove()
+
+
+@pytest.mark.vcr
+@pytest.mark.scihub
+def test_download_all_quicklooks(api, tmpdir, quicklook_products):
+    ids = [product["id"] for product in quicklook_products]
+
+    # Download normally
+    downloaded_quicklooks, failed_quicklooks = api.download_all_quicklooks(
+        ids, str(tmpdir), n_concurrent_dl=1
+    )
+    assert len(failed_quicklooks) == 0
+    assert len(downloaded_quicklooks) == len(ids)
+    for product_id, product_info in downloaded_quicklooks.items():
+        pypath = py.path.local(product_info["path"])
+        assert pypath.check(exists=1, file=1)
+        assert pypath.purebasename in product_info["title"]
+        assert pypath.size() == product_info["quicklook_size"]
+
+    tmpdir.remove()
+
+
+@pytest.mark.vcr
+@pytest.mark.scihub
+def test_download_all_quicklooks_one_fail(api, tmpdir, quicklook_products):
+    ids = [product["id"] for product in quicklook_products]
+
+    # Force one download to fail
+    id = ids[0]
+    with requests_mock.mock(real_http=True) as rqst:
+        url = "https://scihub.copernicus.eu/apihub/odata/v1/Products('{id}')/Products('Quicklook')/$value".format(
+            id=id
+        )
+        headers = api.session.get(url).headers
+        headers["content-type"] = "image/xxxx"
+        rqst.get(url, headers=headers)
+        downloaded_quicklooks, failed_quicklooks = api.download_all_quicklooks(
+            ids, str(tmpdir), n_concurrent_dl=1
+        )
+        assert len(failed_quicklooks) == 1
+        assert len(downloaded_quicklooks) + len(failed_quicklooks) == len(ids)
+        assert id in failed_quicklooks
+
+    tmpdir.remove()
+
+
+@pytest.mark.vcr
+@pytest.mark.scihub
+def test_download_quicklook_invalid_id(api):
+    uuid = "1f62a176-c980-41dc-xxxx-c735d660c910"
+    with pytest.raises(SentinelAPIError) as excinfo:
+        api.download_quicklook(uuid)
+    assert "Invalid key" in excinfo.value.msg